Refactor functions in DataAquisition.py
I refactored the function 'get_single_school_data()' into the following
functions:
    - get_single_school_data()
    - get_single_school_absence_and_pupil_data()
    - get_single_school_absence_and_pupil_url()
    - get_single_school_primary_data()
    - get_single_school_primary_url()
    - get_soup()

I made the choice to refactor the function 'get_single_school_data()'
to reduce the functionality of each individual function. I modified the
documentation for the class to reflect this change. I also created the
function header for each of the new functions above, including the
documentation for each function.

I decided to remove the optional parameter 'data_rows' from the class
since I decided that it is best for all of the relevant data to be
obtained by default (rather than allowing the user to choose which data
they would like to scrape).

I modified the documentation for the class to include a list of all the
data columns that will be provided for each primary school in the UK.

I created a TODO list of additional data columns that could be scraped
from the gov.uk website. I have decided not to include these data
columns to limit the initial scope of the project.