Modify documentation for DataAquisition.py
Modified the documentation for the script DataAquisition.py to include
all of the absence and pupil information that will be scraped from the
gov.uk website.

Created the mock pd.DataFrame which contains the absence and pupil
information that should be returned by the function 'get_single_...
...school_absence_and_pupil_data' when given the parameters
'St Anne's Catholic Primary School, Streetly' and '104241'. This mock
data will be used in the test_DataAquisition class.