Create new unit test case in test_DataAquisition
Create the unit test case 'test_get_single_school_data_correct_return'
which tests that the function 'get_single_school_data' returns the
correct pd.DataFrame, which should contain all of the data described
in the documentation for the 'DataAquisition.py' script. This is tested
for the school 'St Anne's Catholic Primary School, Streetly' which has
the URN '104241'.

Create the .csv file 'mock_get_single_school_data_test.csv' which stores
the mock pd.DataFrame which contains the pd.DataFrame that is expected
to be returned by the function 'get_single_school_data' when called with
the two parameters 'St Anne's Catholic Primary School, Streetly' and
'104241'.