Modify assert sections of unit tests
Modified the constant EXPECTED_PARLIAMENTARY_CONSTITUENT_LIST to make
it a set rather than a list. This is because the order of the elements
returned by the functions in the DataAquisition module is not important.

Modified the assert sections of the following unit tests, casting the
returned lists to a set before comparing them to the expected set:
- test_read_parliamentary_constituencies_correct_return
- test_get_parliamentary_constituencies_file_exists_correct_return
- test_get_parliamentary_constituencies_file_does_not_exist_correct...
..._return
- test_scrape_parliamentary_constituencies_correct_return