Write function get_user_agent in DataAcquisition
Wrote the function get_user_agent in DataAcquisition.py which returns
the user agent to be used in the requests. This user agent is stored
in the file user_agent.txt and is read from there. If the file does not
exist then a FileNotFoundError is raised.

The tests below now pass:
- test_get_user_agent_file_exists_correct_return
- test_get_user_agent_file_does_not_exist_correct_output

Modified the assert section of the unit test test_get_user_agent_file...
..._does_not_exist_correct_output to include the expected output as
a constant which is defined at the start of the test file.