Edit test data file to reduce the amount of data
I modified the test data file
mock_uk_parliamentary_constituencies_test.txt to only include the
parliamentary constituencies in England. I made this change since I
decided that I only want to analyse schools in England and not the whole
of the UK.

I modified the docstring of the test class TestDataAcquisition to
reflect the change in the data file
mock_uk_parliamentary_constituencies_test.txt.